Refactoring arg extractors - first pass
