+++ cas-server-core/src/main/java/org/jasig/cas/authentication/principal/SimpleWebApplicationServiceImpl.java
@@ -18,13 +18,12 @@
  */
 package org.jasig.cas.authentication.principal;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
-
+import org.jasig.cas.CasProtocolConstants;
 import org.jasig.cas.authentication.principal.Response.ResponseType;
 import org.springframework.util.StringUtils;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Represents a service which wishes to use the CAS protocol.
@@ -33,14 +32,6 @@ import org.springframework.util.StringUtils;
  * @since 3.1
  */
 public final class SimpleWebApplicationServiceImpl extends AbstractWebApplicationService {
-
-    private static final String CONST_PARAM_SERVICE = "service";
-
-    private static final String CONST_PARAM_TARGET_SERVICE = "targetService";
-
-    private static final String CONST_PARAM_TICKET = "ticket";
-
-    private static final String CONST_PARAM_METHOD = "method";
 
     private final ResponseType responseType;
 
@@ -63,45 +54,20 @@ public final class SimpleWebApplicationServiceImpl extends AbstractWebApplicatio
      * @param artifactId the artifact id
      * @param responseType the response type
      */
-    private SimpleWebApplicationServiceImpl(final String id,
+    public SimpleWebApplicationServiceImpl(final String id,
         final String originalUrl, final String artifactId,
         final ResponseType responseType) {
         super(id, originalUrl, artifactId);
         this.responseType = responseType;
     }
 
-    /**
-     * Creates the service from the request.
-     *
-     * @param request the request
-     * @return the simple web application service impl
-     */
-    public static SimpleWebApplicationServiceImpl createServiceFrom(
-        final HttpServletRequest request) {
-        final String targetService = request
-            .getParameter(CONST_PARAM_TARGET_SERVICE);
-        final String method = request.getParameter(CONST_PARAM_METHOD);
-        final String serviceToUse = StringUtils.hasText(targetService)
-            ? targetService : request.getParameter(CONST_PARAM_SERVICE);
-
-        if (!StringUtils.hasText(serviceToUse)) {
-            return null;
-        }
-
-        final String id = cleanupUrl(serviceToUse);
-        final String artifactId = request.getParameter(CONST_PARAM_TICKET);
-
-        return new SimpleWebApplicationServiceImpl(id, serviceToUse,
-            artifactId, "POST".equals(method) ? ResponseType.POST
-                : ResponseType.REDIRECT);
-    }
 
     @Override
     public Response getResponse(final String ticketId) {
         final Map<String, String> parameters = new HashMap<String, String>();
 
         if (StringUtils.hasText(ticketId)) {
-            parameters.put(CONST_PARAM_TICKET, ticketId);
+            parameters.put(CasProtocolConstants.PARAM_TICKET, ticketId);
         }
 
         if (ResponseType.POST == this.responseType) {
