+++ cas-server-core/src/main/java/org/jasig/cas/web/support/AbstractArgumentExtractor.java
@@ -18,11 +18,11 @@
  */
 package org.jasig.cas.web.support;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.jasig.cas.authentication.principal.WebApplicationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * Abstract class for handling argument extraction.
@@ -49,6 +49,33 @@ public abstract class AbstractArgumentExtractor implements ArgumentExtractor {
         }
 
         return service;
+    }
+
+    /**
+     * Cleanup the url. Removes jsession ids and query strings.
+     *
+     * @param url the url
+     * @return sanitized url.
+     */
+    protected String cleanupUrl(final String url) {
+        if (url == null) {
+            return null;
+        }
+
+        final int jsessionPosition = url.indexOf(";jsession");
+
+        if (jsessionPosition == -1) {
+            return url;
+        }
+
+        final int questionMarkPosition = url.indexOf("?");
+
+        if (questionMarkPosition < jsessionPosition) {
+            return url.substring(0, url.indexOf(";jsession"));
+        }
+
+        return url.substring(0, jsessionPosition)
+                + url.substring(questionMarkPosition);
     }
 
     /**
