+++ cas-server-support-openid/src/main/java/org/jasig/cas/support/openid/OpenIdProtocolConstants.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed to Jasig under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work
+ * for additional information regarding copyright ownership.
+ * Jasig licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License.  You may obtain a
+ * copy of the License at the following location:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.jasig.cas.support.openid;
+
+/**
+ * Class that exposes relevant constants and parameters to
+ * the OpenId protocol. These include attribute names, pre-defined
+ * values and expected request parameter names as is specified
+ * by the protocol.
+ *
+ * @author Misagh Moayyed
+ * @since 4.1
+ */
+public interface OpenIdProtocolConstants {
+    /** Constant representing the openid return service. */
+    String PARAM_SERVICE = "openid.return_to";
+}
