+++ cas-server-support-openid/src/main/java/org/jasig/cas/support/openid/web/support/OpenIdArgumentExtractor.java
@@ -18,11 +18,14 @@
  */
 package org.jasig.cas.support.openid.web.support;
 
-import javax.servlet.http.HttpServletRequest;
-
 import org.jasig.cas.authentication.principal.WebApplicationService;
+import org.jasig.cas.support.openid.OpenIdProtocolConstants;
 import org.jasig.cas.support.openid.authentication.principal.OpenIdService;
-import org.jasig.cas.web.support.ArgumentExtractor;
+import org.jasig.cas.web.support.AbstractArgumentExtractor;
+import org.openid4java.message.ParameterList;
+import org.springframework.util.StringUtils;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * Constructs an OpenId Service.
@@ -30,10 +33,33 @@ import org.jasig.cas.web.support.ArgumentExtractor;
  * @author Scott Battaglia
  * @since 3.1
  */
-public class OpenIdArgumentExtractor implements ArgumentExtractor {
+public class OpenIdArgumentExtractor extends AbstractArgumentExtractor {
 
     @Override
-    public WebApplicationService extractService(final HttpServletRequest request) {
-        return OpenIdService.createServiceFrom(request);
+    public WebApplicationService extractServiceInternal(final HttpServletRequest request) {
+        return createServiceFrom(request);
+    }
+
+    /**
+     * Creates the service from the request.
+     *
+     * @param request the request
+     * @return the OpenID service
+     */
+    protected OpenIdService createServiceFrom(final HttpServletRequest request) {
+        final String service = request.getParameter(OpenIdProtocolConstants.PARAM_SERVICE);
+        final String openIdIdentity = request.getParameter("openid.identity");
+        final String signature = request.getParameter("openid.sig");
+
+        if (openIdIdentity == null || !StringUtils.hasText(service)) {
+            return null;
+        }
+
+        final String id = cleanupUrl(service);
+        final String artifactId = request.getParameter("openid.assoc_handle");
+        final ParameterList paramList = new ParameterList(request.getParameterMap());
+
+        return new OpenIdService(id, service, artifactId, openIdIdentity,
+                signature, paramList);
     }
 }
