+++ cas-server-support-saml/src/main/java/org/jasig/cas/support/saml/web/support/GoogleAccountsArgumentExtractor.java
@@ -18,17 +18,29 @@
  */
 package org.jasig.cas.support.saml.web.support;
 
+import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.io.IOUtils;
 import org.jasig.cas.authentication.principal.WebApplicationService;
 import org.jasig.cas.services.ServicesManager;
+import org.jasig.cas.support.saml.SamlProtocolConstants;
 import org.jasig.cas.support.saml.authentication.principal.GoogleAccountsService;
+import org.jasig.cas.support.saml.util.SamlUtils;
 import org.jasig.cas.web.support.AbstractArgumentExtractor;
+import org.jdom.Document;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.util.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.constraints.NotNull;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.UnsupportedEncodingException;
 import java.security.PrivateKey;
 import java.security.PublicKey;
+import java.util.zip.DataFormatException;
+import java.util.zip.Inflater;
+import java.util.zip.InflaterInputStream;
 
 /**
  * Constructs a GoogleAccounts compatible service and provides the public and
@@ -52,8 +64,7 @@ public final class GoogleAccountsArgumentExtractor extends AbstractArgumentExtra
     
     @Override
     public WebApplicationService extractServiceInternal(final HttpServletRequest request) {
-        return GoogleAccountsService.createServiceFrom(request,
-                this.privateKey, this.publicKey, this.servicesManager);
+        return createServiceFrom(request, this.privateKey, this.publicKey, this.servicesManager);
     }
 
     /**
@@ -100,5 +111,139 @@ public final class GoogleAccountsArgumentExtractor extends AbstractArgumentExtra
     @Deprecated
     public void setAlternateUsername(final String alternateUsername) {
         LOGGER.warn("setAlternateUsername() is deprecated and has no effect. Instead use the configuration in service registry.");
+    }
+
+    /**
+     * Creates the service from request.
+     *
+     * @param request the request
+     * @param privateKey the private key
+     * @param publicKey the public key
+     * @param servicesManager the services manager
+     * @return the google accounts service
+     */
+    protected GoogleAccountsService createServiceFrom(
+            final HttpServletRequest request, final PrivateKey privateKey,
+            final PublicKey publicKey, final ServicesManager servicesManager) {
+        final String relayState = request.getParameter(SamlProtocolConstants.CONST_RELAY_STATE);
+
+        final String xmlRequest = decodeAuthnRequestXML(request.getParameter(SamlProtocolConstants.CONST_PARAM_SERVICE));
+
+        if (!StringUtils.hasText(xmlRequest)) {
+            return null;
+        }
+
+        final Document document = SamlUtils.constructDocumentFromXmlString(xmlRequest);
+
+        if (document == null) {
+            return null;
+        }
+
+        final String assertionConsumerServiceUrl = document.getRootElement().getAttributeValue("AssertionConsumerServiceURL");
+        final String requestId = document.getRootElement().getAttributeValue("ID");
+
+        return new GoogleAccountsService(assertionConsumerServiceUrl,
+                relayState, requestId, privateKey, publicKey, servicesManager);
+    }
+
+    /**
+     * Deflate the given bytes using zlib.
+     *
+     * @param bytes the bytes
+     * @return the converted string
+     */
+    private String zlibDeflate(final byte[] bytes) {
+        final ByteArrayInputStream bais = new ByteArrayInputStream(bytes);
+        final ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        final InflaterInputStream iis = new InflaterInputStream(bais);
+        final byte[] buf = new byte[1024];
+
+        try {
+            int count = iis.read(buf);
+            while (count != -1) {
+                baos.write(buf, 0, count);
+                count = iis.read(buf);
+            }
+            return new String(baos.toByteArray());
+        } catch (final Exception e) {
+            return null;
+        } finally {
+            IOUtils.closeQuietly(iis);
+        }
+    }
+
+    /**
+     * Base64 decode.
+     *
+     * @param xml the xml
+     * @return the byte[]
+     */
+    private byte[] base64Decode(final String xml) {
+        try {
+            final byte[] xmlBytes = xml.getBytes("UTF-8");
+            return Base64.decodeBase64(xmlBytes);
+        } catch (final Exception e) {
+            return null;
+        }
+    }
+
+    /**
+     * Decode authn request xml.
+     *
+     * @param encodedRequestXmlString the encoded request xml string
+     * @return the request
+     */
+    private String decodeAuthnRequestXML(
+            final String encodedRequestXmlString) {
+        if (encodedRequestXmlString == null) {
+            return null;
+        }
+
+        final byte[] decodedBytes = base64Decode(encodedRequestXmlString);
+
+        if (decodedBytes == null) {
+            return null;
+        }
+
+        final String inflated = inflate(decodedBytes);
+
+        if (inflated != null) {
+            return inflated;
+        }
+
+        return zlibDeflate(decodedBytes);
+    }
+
+    /**
+     * Inflate the given byte array.
+     *
+     * @param bytes the bytes
+     * @return the string
+     */
+    private static String inflate(final byte[] bytes) {
+        final Inflater inflater = new Inflater(true);
+        final byte[] xmlMessageBytes = new byte[10000];
+
+        final byte[] extendedBytes = new byte[bytes.length + 1];
+        System.arraycopy(bytes, 0, extendedBytes, 0, bytes.length);
+        extendedBytes[bytes.length] = 0;
+
+        inflater.setInput(extendedBytes);
+
+        try {
+            final int resultLength = inflater.inflate(xmlMessageBytes);
+            inflater.end();
+
+            if (!inflater.finished()) {
+                throw new RuntimeException("buffer not large enough.");
+            }
+
+            inflater.end();
+            return new String(xmlMessageBytes, 0, resultLength, "UTF-8");
+        } catch (final DataFormatException e) {
+            return null;
+        } catch (final UnsupportedEncodingException e) {
+            throw new RuntimeException("Cannot find encoding: UTF-8", e);
+        }
     }
 }
